requirements.txt
django
django-ckeditor
pillow
djangorestframework